some tweks, still buggy and strange behaviour

+++ srcs/vm/funcs/0x09_zjmp.c
@@ -6,7 +6,7 @@
 /*   By: gmolin <gmolin@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/06 15:02:48 by seronen           #+#    #+#             */
-/*   Updated: 2021/02/13 18:53:15 by gmolin           ###   ########.fr       */
+/*   Updated: 2021/02/13 19:45:26 by gmolin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ void    op_zjmp(t_carriage *carriage, unsigned char *arena)
 	if (carriage->carry_flag)
 	{
 		// arg_idx_value = carriage->stmt->args[0] % IDX_MOD;
-        arg_idx_value = real_modulo(carriage->abs_pos, (short)carriage->stmt->args[0], IDX_MOD);
+        arg_idx_value = real_modulo(carriage->abs_pos, carriage->stmt->args[0], IDX_MOD);
 		carriage->pos = &arena[arg_idx_value];
 		carriage->abs_pos = carriage->abs_pos + arg_idx_value - 3;
 	}
